requirements.txt
chromadb~=1.0.11
dotenv~=0.9.9
google-genai~=1.17.0
google-generativeai~=0.8.5
numpy~=2.2.6
pyprojroot~=0.3.0
python-box~=7.3.2
requests~=2.32.3
scikit-learn~=1.6.1
streamlit~=1.45.1
PyYAML~=6.0.2
python-dotenv~=1.1.0
protobuf~=5.29.5
tiktoken~=0.9.0